fix missing service suffix in packagename
