+++ wls-basisdaten-service/src/main/java/de/muenchen/oss/wahllokalsystem/basisdatenservice/security/AuthUtils.java
@@ -1,0 +1,41 @@
+/*
+ * Copyright (c): it@M - Dienstleister für Informations- und Telekommunikationstechnik
+ * der Landeshauptstadt München, 2024
+ */
+package de.muenchen.oss.wahllokalsystem.basisdatenservice.security;
+
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.oauth2.server.resource.authentication.JwtAuthenticationToken;
+
+/**
+ * Utilities zu Authentifizierungsdaten.
+ */
+public class AuthUtils {
+
+    public static final String NAME_UNAUTHENTICATED_USER = "unauthenticated";
+
+    private static final String TOKEN_USER_NAME = "user_name";
+
+    private AuthUtils() {
+    }
+
+    /**
+     * Extrahiert den Usernamen aus dem vorliegenden Spring Security Context via
+     * {@link SecurityContextHolder}.
+     *
+     * @return der Username or a "unauthenticated", wenn keine {@link Authentication} existiert
+     */
+    public static String getUsername() {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (authentication instanceof JwtAuthenticationToken jwtAuthenticationToken) {
+            return (String) jwtAuthenticationToken.getTokenAttributes().getOrDefault(TOKEN_USER_NAME, null);
+        } else if (authentication instanceof UsernamePasswordAuthenticationToken usernamePasswordAuthenticationToken) {
+            return usernamePasswordAuthenticationToken.getName();
+        } else {
+            return NAME_UNAUTHENTICATED_USER;
+        }
+    }
+
+}
