--- wls-basisdaten-service/src/test/java/de/muenchen/oss/wahllokalsystem/basisdaten/rest/TheEntityRepositoryTest.java
@@ -1,55 +1,0 @@
-/*
- * Copyright (c): it@M - Dienstleister für Informations- und Telekommunikationstechnik
- * der Landeshauptstadt München, 2024
- */
-package de.muenchen.oss.wahllokalsystem.basisdaten.rest;
-
-import static de.muenchen.oss.wahllokalsystem.basisdaten.TestConstants.SPRING_NO_SECURITY_PROFILE;
-import static de.muenchen.oss.wahllokalsystem.basisdaten.TestConstants.SPRING_TEST_PROFILE;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNotNull;
-
-import de.muenchen.oss.wahllokalsystem.basisdaten.MicroServiceApplication;
-import de.muenchen.oss.wahllokalsystem.basisdaten.domain.TheEntity;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.ActiveProfiles;
-import org.springframework.transaction.annotation.Propagation;
-import org.springframework.transaction.annotation.Transactional;
-
-@SpringBootTest(
-        classes = { MicroServiceApplication.class },
-        webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT,
-        properties = {
-                "spring.datasource.url=jdbc:h2:mem:wls;DB_CLOSE_ON_EXIT=FALSE",
-                "refarch.gracefulshutdown.pre-wait-seconds=0"
-        }
-)
-@ActiveProfiles(profiles = { SPRING_TEST_PROFILE, SPRING_NO_SECURITY_PROFILE })
-class TheEntityRepositoryTest {
-
-    @Autowired
-    private TheEntityRepository repository;
-
-    @Test
-    @Transactional(propagation = Propagation.REQUIRED, noRollbackFor = Exception.class)
-    void testSave() {
-
-        // Implement your logic here by replacing and/or extending the code
-
-        // initialize
-        TheEntity original = new TheEntity();
-        original.setTextAttribute("test");
-
-        // persist
-        original = repository.save(original);
-
-        // check
-        TheEntity persisted = repository.findById(original.getId()).orElse(null);
-        assertNotNull(persisted);
-        assertEquals(original, persisted);
-
-    }
-
-}
